code split (#2024)
* :sparkles: PRTL-1457 code split by route

* split on common components
