+++ package.json
@@ -37,6 +37,7 @@
     "filesaver.js": "0.2.0",
     "filesize": "3.3.0",
     "glamor": "2.20.25",
+    "import-inspector": "2.0.0",
     "innersvg-polyfill": "0.0.2",
     "js-cookie": "2.1.4",
     "jsurl": "0.1.5",
@@ -55,6 +56,7 @@
     "react-helmet": "5.1.3",
     "react-highlighter": "0.3.3",
     "react-icons": "2.2.5",
+    "react-loadable": "4.0.2",
     "react-measure": "1.4.5",
     "react-modal": "1.7.7",
     "react-progress": "0.0.12",
@@ -73,6 +75,7 @@
     "url-join": "1.1.0"
   },
   "devDependencies": {
+    "babel-plugin-import-inspector": "2.0.0",
     "babel-plugin-relay": "1.0.1",
     "babel-plugin-transform-export-extensions": "6.22.0",
     "eslint-config-prettier": "2.1.1",
