+++ src/packages/@ncigdc/components/LoadableWithLoading/index.js
@@ -1,0 +1,10 @@
+import Loadable from 'react-loadable';
+import Loading from './Loading';
+
+const LoadableWithLoading = options =>
+  Loadable({
+    loading: Loading,
+    ...options,
+  });
+
+export default LoadableWithLoading;
