+++ src/packages/@ncigdc/modern_components/CancerDistributionTable/index.js
@@ -1,4 +1,9 @@
-import Component from './CancerDistributionTable';
+import LoadableWithLoading from '@ncigdc/components/LoadableWithLoading';
 import createRenderer from './CancerDistributionTable.relay';
 import createRendererProjects from './Projects.relay';
-export default createRenderer(createRendererProjects(Component));
+
+const AsyncComponent = LoadableWithLoading({
+  loader: () => import('./CancerDistributionTable'),
+});
+
+export default createRenderer(createRendererProjects(AsyncComponent));
