+++ src/packages/@ncigdc/modern_components/GeneSymbol/index.js
@@ -1,6 +1,11 @@
+import LoadableWithLoading from '@ncigdc/components/LoadableWithLoading';
 import PropTypes from 'prop-types';
-import Component from './GeneSymbol';
 import createRenderer from './GeneSymbol.relay';
+
+const Component = LoadableWithLoading({
+  loader: () => import('./GeneSymbol'),
+});
+
 const GeneSymbol = createRenderer(Component);
 GeneSymbol.propTypes = {
   geneId: PropTypes.string,
