+++ src/packages/@ncigdc/modern_components/GenesBarChart/index.js
@@ -1,2 +1,6 @@
-import GenesBarChart from './GenesBarChart';
+import LoadableWithLoading from '@ncigdc/components/LoadableWithLoading';
+
+const GenesBarChart = LoadableWithLoading({
+  loader: () => import('./GenesBarChart'),
+});
 export default GenesBarChart;
